Refactor deliveryservice API minor versioning (#3713)
* Refactor deliveryservice API minor versioning

Remove all version-specific logic from the deliveryservices API
implementation except for the logic of upgrading requests of previous
minor versions into requests of the latest minor version to be handled
as such. All main logic now uses and belongs to the "latest" struct, so
it is no longer confusing as to which methods need updated/called for
various minor version structs. The versioned structs are now only
required and used for unmarshalling requests of a specific minor
version, upgrading those requests to the latest minor version, and
marshalling responses back down into the requested version. It should be
much more clear now how to implement and add a new minor version to the
deliveryservices API (also with updated README instructions).

* Address review comments and remove a couple redundant functions